refactor(status): generalize the section-registry pattern to all of collect()
collect() (CC 250, the worst function in the repo) inlined ~25 subsystem
concerns as ad-hoc try/catch + branch ladders. The file already had the
right pattern for exactly one concern (HOST_DETAIL_RENDERERS +
renderHostDetailRows); this generalizes it to the rest.

Each concern becomes its own module under src/commands/status/sections/,
exporting { id, collect: async (ctx) => Row[] } with
ctx = { cfg, cwd, pkgRoot, integrationFacts }. collect() is now two ordered
walks over SECTIONS_BEFORE_HOST_DETAIL / SECTIONS_AFTER_HOST_DETAIL (split
only because three existing calls -- collectDejaVuRows,
renderHostDetailRows, admittedLifecycleFallbackRows -- keep their own
bespoke signatures and error contracts between them, unchanged), each row
wrapped in a uniform try/catch that falls back to a generic
'<id> check unavailable' warn row. Sections that already had their own
try/catch (most of them) keep it verbatim for their exact original
message; the uniform wrapper is a backstop, and for the handful of
concerns that had NO try/catch before (security, learning, aqe, agentdb,
mcp, statusline, qe-court), it's a strict improvement: an unexpected throw
there now degrades one row instead of crashing all of collect().

The providers section (~8 sub-concerns under one try/catch, per audit) is
split into five sections -- providers-status, providers-external-intent,
providers-external-projection, providers-ruflo-models,
providers-local-bindings -- sharing a small computeProviderExternalState
helper (_providers-external.mjs) that each calls and catches
independently, so one probe failing no longer collapses all eight rows
into a single warn. Its drift-comparison logic duplicates write-side logic
in src/lib/providers.mjs by design for now; carries a
"TODO(complexity-program)" marker for a later cross-track re-homing.
The three-block codex-mcp concern and the four-block statusline concern
each become one section file with independently-caught inner functions,
preserving their existing per-probe error isolation.

Result: collect() CC 250 -> ~5 (a loop + a try/catch), status.mjs
1216 -> 118 lines. Every new section under CC 25 (worst is 21). Pure
decomposition: no messages, row order, or logic changed -- verified
byte-for-byte against the golden snapshot and all 48 status-command
behavior tests, plus the full status-aqe-drift and status-viability
suites.
